job-manager: relocate submit handler
Move the submit request handler out of job-manager.c
(the mainline) to its own source, submit.c.  This
aligns its structure with other request handlers
in the job manager.

The code itself didn't change substantively.
